requirements.txt
PyYAML==6.0.1
boto3==1.34.40
watchtower==3.0.1
pymongo==4.6.1
requests==2.31.0